remove legacy traits, inline parsers inside the HttpHeaderParser; improve Accept-Charset, Accept-Encoding, Accept-Language parsers
